+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Data Transfer Object classes generation from swagger spec
 - Relationships support in DTO classes
+- Factory and Middleware Decorator for assign middleware to methods
 
 ## [0.0.6] - 2020-05-08
 ### Added